+++ js/accordion/Index.svelte
@@ -2,15 +2,24 @@
 	import Accordion from "./shared/Accordion.svelte";
 	import { Block } from "@gradio/atoms";
 	import { StatusTracker } from "@gradio/statustracker";
-	import type { LoadingStatus } from "@gradio/statustracker";
 
 	import { BaseColumn } from "@gradio/column";
 	import { Gradio } from "@gradio/utils";
+	import type { SharedProps } from "@gradio/utils";
 
 	import type { AccordionProps, AccordionEvents } from "./types";
 
 	let props = $props();
-	const gradio = new Gradio<AccordionEvents, AccordionProps>(props);
+	class AccordionGradio extends Gradio<AccordionEvents, AccordionProps> {
+		set_data(data: Partial<object & SharedProps>): void {
+			if ("open" in data && data.open) {
+				this.dispatch("gradio_expand");
+			}
+			super.set_data(data);
+			this.shared.loading_status.status = "complete";
+		}
+	}
+	const gradio = new AccordionGradio(props);
 
 	let label = $derived(gradio.shared.label || "");
 </script>