+++ js/_spaces-test/CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ### Dependency updates
 
+- @gradio/client@2.0.4
+
+## 0.0.1
+
+### Dependency updates
+
 - @gradio/client@2.0.3
 - @gradio/theme@0.6.1